Develop

+++ src/main/java/com/demo/service/UserService.java
@@ -3,8 +3,10 @@ package com.demo.service;
 import com.demo.model.CurrencyType;
 import com.demo.model.User;
 
+import java.math.BigDecimal;
+
 public interface UserService {
     User findUserByEmail(String email);
     void saveUser(User user);
-    Double getBalanceByCurrencyType(User user, CurrencyType type);
+    BigDecimal getBalanceByCurrencyType(User user, CurrencyType type);
 }
